2단계: 내 예약 조회 API 응답에 결제정보 추가

+++ src/main/java/roomescape/payment/Payment.java
@@ -16,7 +16,7 @@ public class Payment {
     private String paymentKey;
     private String orderId;
     private String orderName;
-    private int totalAmount;
+    private long totalAmount;
     private String method;
     private String status;
     private LocalDateTime requestedAt;
